packet: separate bmp stuff from bgp.go
